eventfy-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
supabase==2.11.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.19
Pillow==11.1.0
qrcode[pil]==8.0
reportlab==4.2.5
httpx==0.28.1
python-dotenv==1.0.1
